refactor: refactor redux file structure imports
