+++ package.json
@@ -35,6 +35,7 @@
     "@types/react-native": "^0.63.48",
     "@types/react-redux": "^7.1.16",
     "@types/react-test-renderer": "^17.0.0",
+    "@types/redux": "^3.6.0",
     "@types/redux-mock-store": "^1.0.2",
     "@typescript-eslint/eslint-plugin": "^4.15.0",
     "@typescript-eslint/parser": "^4.15.0",
