+++ src/actions/countActions.ts
@@ -1,8 +1,8 @@
-import { ADD_NUMBER } from '../constants/actionConstants';
+import { ADD_NUMBER } from '../constants';
 
 // TODO: Fix the type in Future PR
 // eslint-disable-next-line import/prefer-default-export
 export const add = (num: number): any => ({
   type: ADD_NUMBER,
-  payLoad: num,
+  payload: num,
 });
