+++ src/constants/authConstants.ts
@@ -1,3 +1,2 @@
-export const ADD_NUMBER = 'ADD_NUMBER';
 export const LOG_IN = 'AUTH/LOGIN';
 export const LOG_OUT = 'AUTH/LOGOUT';
