+++ src/constants/countConstants.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const ADD_NUMBER = 'ADD_NUMBER';
